Refactor ViewModel and related imports to improve code organization
